IValue is part of the overall runtime system, not just the JIT. So it
should be tested in the ATen tests.

The real motivation though is so that I can use gtest directly, not the
hacked-up version the JIT uses.

[ghstack-poisoned]

+++ caffe2/operators/fused_rowwise_nbitfake_conversion_ops.h
@@ -80,7 +80,7 @@ class FloatToFusedNBitFakeRowwiseQuantizedOp final
 #endif
 
 #pragma omp parallel for if (GREEDY)
-    for (size_t row = 0; row < input_rows; ++row) {
+    for (int row = 0; row < input_rows; ++row) {
       float* tmp = tmp_vec.data();
 #ifdef _OPENMP
       if (GREEDY) {
